Done a new feature where allow global admins change the institution of a space
Before this feature, global administrators were unable to change the institution of a space after its creation. Global administrators should be able to change the institution of a space. So, to do that, an option with all insitutions was inserted into the edit page of a space (just visible to global administrators), and a guard function was created into the model to guarantee this action.

Fixes #1069