Done a new feature for limit the number of users that can record in an institution
Before this feature, institutions don't had a way to limit number of users with the ability to record meetings. To do that I created a new field inside the institution database, and all functions needed to validate the quantity of users with this capability inside each institution. Now if any institution reach the limit, administrators can't set the record ability to other members until some member with the ability is removed.

Fixes #1081