Fixed a mistake related to setting the record meetings ability
After some tests was detected a problem related to this change, so to fix it was changed the idea of the condition. Now global administrators can set the record meetings ability to any user even if the limit of users who can record in a institution are over the limit.

Refs #1081