Improved UsersController#update and added some tests to it
Modified the update condition concerning the can_record variable. This was needed because the test showed some problems when a normal user try to update this variable. Added a few more tests for it. Now global administrators, institutional administrators, and normal users are tested to ensure the update was ok when modified the can_record variable.


Refs #1081