Fix user/space losing institution on save
When the setting of an institution in a user/space was changed from setting
by name to setting by id, the select2 components weren't properly adapted.
Had to fix the way the first element is set in select2 so it knows how to
show the institution name but submit the institution id.

refs #1250