Change and fix how institution secrets interact with bbb_rails
Now the application overrides the "server" methods in the models that
need a server (rooms, meetings and recordings) to return a server with
the secret associated with the resource's institution. The server returned
is only stored in memory.

Overrides also the full synchronization update of recordings to also send
getRecordings using the secrets registered in all institutions.

This is a working solution that fixes the update of recordings for
institutions with secrets.